Found bug
I didn't spell the name of a collection in firestore correctly. Go figure. Now it works correctly, users can add sets to each exercise and each workout is unique to a specific date. Sets also are listed in the order they were created (i.e. newer sets are at the bottom of a list)