Fixed bug / created new function in firebase.js
Fixed bug in NewEx.jsx that caused an issue when adding new exercise while in selected category page. I just had to add some conditional logic to form state. Also started the boilerplate of a function in firebase.js that will allow users to use the search tool for exercises.