Update to CurrentWorkoutList.jsx, index.css
Added a check to see if ex name is too long, will slice the end and add an ellipsis. This should only work for smaller screens, so I'll need to adjust for larger screens. Also finished the div UI for the edit/delete div for the exercise. Now to finish fixing the functionality.