Changes to Firebase.js
The date was not being passed around correctly. Before, when viewing exercise detail history, it would show the date the exercise was added, even if the user added the exercise to an older date, which causes discrepancies. Now it correctly shows the workout date for the day the exercise was performed.